fix: Delete previous Search component and move logic to a context
